Cleaned up old files
